refactor: Access Token 재발급 관련 테스트 코드 분리
